Consolidate TEI provider methods
